Added all changes needed for final intcode computer!

+++ src/main/java/net/chewett/adventofcode2019/intcode/Intcode.java
@@ -2,26 +2,26 @@ package net.chewett.adventofcode2019.intcode;
 
 public class Intcode {
 
-    private int[] intsFromIntcode;
+    private long[] intsFromIntcode;
 
     public Intcode(String intcode) {
         String[] intcodeParts = intcode.split(",");
-        this.intsFromIntcode = new int[intcodeParts.length];
+        this.intsFromIntcode = new long[intcodeParts.length];
 
         for(int i = 0; i < intcodeParts.length; i++) {
-            this.intsFromIntcode[i] = Integer.parseInt(intcodeParts[i]);
+            this.intsFromIntcode[i] = Long.parseLong(intcodeParts[i]);
         }
     }
 
-    public Intcode(int[] intcode) {
+    public Intcode(long[] intcode) {
         this.intsFromIntcode = intcode;
     }
 
-    public int[] readIntoMemory() {
+    public long[] readIntoMemory() {
         return this.intsFromIntcode.clone();
     }
 
-    public void setIntToAddress(int address, int value) {
+    public void setIntToAddress(int address, long value) {
         this.intsFromIntcode[address] = value;
     }
