+++ src/main/java/net/chewett/adventofcode2019/intcode/instructions/FinishInstruction.java
@@ -15,7 +15,7 @@ public class FinishInstruction extends IntcodeInstruction {
     }
 
     @Override
-    public int getIntcodeInstructionNumber() {
+    public long getIntcodeInstructionNumber() {
         return 99;
     }
